feat: Minor improvements
