+++ components/src/lib.rs
@@ -1,6 +1,7 @@
 //! # Freya Components
 //! A collection of basic components to be used in Freya.
 
+mod accordion;
 mod button;
 mod dropdown;
 mod external_link;
@@ -12,6 +13,7 @@ mod switch;
 mod theme;
 mod tooltip;
 
+pub use accordion::*;
 pub use button::*;
 pub use dropdown::*;
 pub use external_link::*;
